+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 - Added basic [Read Me](README.md).
 - Added basic [Game Design Document](GDD.md).
 - Added basic [Change Log](CHANGELOG.md).
+
+# Version 0.0.1 alpha - 18/02/2026
+- Added folders
+- Added sprites
+- Added basic character class
+- Added 3 skills
+- Added basic game level withsmall combat test